Create Itsme.md
Hello, World!
My name is Pedro Lucas Braga, I'm a 17 years old student from Brazil, looking to expand my knowledge and reach for the stars. Currently, I am a economics student with a passion on development and computer science. My goal It's to somehow integrate economics and development to create something that will actually make a difference in our world. For now, I just focus on learning as much as I can!